mvSuSiE PIP and Effect Plots

Create the PIP plot and accompanying effect plot showing the effect estimates and significance of the effects for all the traits.

mvsusie_plot_simple(
  fit,
  pos = seq(1, length(fit$variable_names)),
  markers = fit$variable_names,
  chr = 1,
  poslim = range(pos),
  conditions,
  lfsr_cutoff = 0.01,
  cs_colors = c("#1f78b4", "#33a02c", "#e31a1c", "#ff7f00", "#6a3d9a", "#b15928",
    "#a6cee3", "#b2df8a", "#fb9a99", "#fdbf6f", "#cab2d6", "#ffff99", "gray", "cyan")
)

Arguments

fit

Describe the fit input here.

pos

Describe the pos input here.

markers

Describe the markers input here.

chr

Describe the chr input here.

poslim

Describe the poslim input here.

conditions

Describe the conditions input here.

lfsr_cutoff

Describe the lfsr_cutoff input here.

cs_colors

Describe the cs_colors input here.

Value

Describe the return value here.

Details

The colors from colorbrewer2.org.