mvSuSiE PIP and Effect Plots

Create the PIP plot and accompanying effect plot showing the effect estimates and significance of the effects for all the traits. A z-scores plot is also created when z-scores are available.

mvsusie_plot(
  fit,
  chr = 1,
  pos = seq(1, length(fit$variable_names)),
  markers = fit$variable_names,
  conditions = fit$condition_names,
  poslim = range(pos),
  lfsr_cutoff = 0.01,
  sentinel_only = TRUE,
  cs_plot = names(fit$sets$cs),
  add_cs = FALSE,
  conditional_effect = TRUE,
  cs_colors = c("#1f78b4", "#33a02c", "#e31a1c", "#ff7f00", "#6a3d9a", "#b15928",
    "#a6cee3", "#b2df8a", "#fb9a99", "#fdbf6f", "#cab2d6", "#ffff99", "gray", "cyan")
)

Arguments

fit

The mvSuSiE fitted model.

chr

The chromosome number.

pos

The positions of the genetic markers. It should have the same length as fit$variable_names.

markers

The names of the genetic markers (usually SNPs).

conditions

The names of the conditions.

poslim

The range of positions to show in the PIP plot.

lfsr_cutoff

The significance level for lfsr. The default is 0.01.

sentinel_only

If TRUE, only plot the sentinel marker for each CS. If FALSE, plot all markers in each CS.

cs_plot

The CSs included in the plot. The default is to show all CSs.

add_cs

If TRUE, add colored dots to the top of the effect plot showing CS membership.

conditional_effect

If TRUE, plot the conditional effect. If FALSE, plot the marginal effect. conditional_effect = TRUE is recommended.

cs_colors

The color palette for CSs.

Value

The output includes the PIP plot, effect plot, z-scores plot (if z-scores are available in fit), and the table of effect estimates at sentinel markers.

Examples

# See the "mvsusie_intro" vignette for examples.