Simulated Multi-trait Fine-Mapping Data Used in Tutorial

Simulated fine-mapping data set used to illustrate mvSuSiE in the tutorial. The data set includes genotype and phenotype data for 574 samples, 1,001 genetic markers and 20 traits. The traits were simulated from the mvSuSiE model with coefficients simdata$B and residual simdata$par$V. This is a simulation with three causal genetic variants at positions 255, 335 and 493; that is, these are the only genetic variants witih nonzero coefficients.

Format

simdata is a list with the following elements:

raw$X

The matrix of simulated genotypes.

raw$Y

The matrix of simulated traits.

Btrue

The coefficients used to simulate the data.

par$V

The residual covariance matrix used to simulated the data.

par$U

The collection of covariance matrices specifying the mvsusie prior.

par$w

The weights associated with the covariance matrices.

sumstats$n

The sample size.

sumstats$LD

The LD computed from raw$X.

sumstats$bhat

The least-squares effect estimates from the single-marker association tests.

sumstats$sehat

The standard errors of the least-squares effect estimates.